+++ concordium-sdk/src/main/java/com/concordium/sdk/ClientV2MapperExtensions.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk;
 
+import com.concordium.grpc.v2.AbsoluteBlockHeight;
 import com.concordium.grpc.v2.AccessStructure;
 import com.concordium.grpc.v2.AccountAddress;
 import com.concordium.grpc.v2.AccountIndex;
@@ -9,6 +10,7 @@ import com.concordium.grpc.v2.BakerId;
 import com.concordium.grpc.v2.BlockItem;
 import com.concordium.grpc.v2.Commitment;
 import com.concordium.grpc.v2.ContractAddress;
+import com.concordium.grpc.v2.Cooldown;
 import com.concordium.grpc.v2.CredentialPublicKeys;
 import com.concordium.grpc.v2.CredentialRegistrationId;
 import com.concordium.grpc.v2.DelegatorInfo;
@@ -21,7 +23,6 @@ import com.concordium.grpc.v2.NextUpdateSequenceNumbers;
 import com.concordium.grpc.v2.Policy;
 import com.concordium.grpc.v2.ProtocolVersion;
 import com.concordium.grpc.v2.ReleaseSchedule;
-import com.concordium.grpc.v2.AbsoluteBlockHeight;
 import com.concordium.grpc.v2.*;
 import com.concordium.sdk.crypto.bulletproof.BulletproofGenerators;
 import com.concordium.sdk.crypto.ed25519.ED25519PublicKey;
@@ -314,6 +315,7 @@ interface ClientV2MapperExtensions {
         builder
                 .Nonce(to(account.getSequenceNumber()))
                 .accountAmount(to(account.getAmount()))
+                .availableBalance(to(account.getAvailableBalance()))
                 .accountReleaseSchedule(to(account.getSchedule()))
                 .accountCredentials(ImmutableMap.copyOf(to(
                         account.getCredsMap(),
@@ -323,7 +325,8 @@ interface ClientV2MapperExtensions {
                 .accountEncryptedAmount(to(account.getEncryptedBalance()))
                 .accountEncryptionKey(ElgamalPublicKey.from(account.getEncryptionKey().getValue().toByteArray()))
                 .accountIndex(to(account.getIndex()))
-                .accountAddress(to(account.getAddress()));
+                .accountAddress(to(account.getAddress()))
+                .cooldowns(copyOf(to(account.getCooldownsList(), ClientV2MapperExtensions::to)));
 
         if (account.hasStake()) {
             switch (account.getStake().getStakingInfoCase()) {
@@ -342,6 +345,14 @@ interface ClientV2MapperExtensions {
 
     static com.concordium.sdk.types.AccountAddress to(AccountAddress address) {
         return com.concordium.sdk.types.AccountAddress.from(address.getValue().toByteArray());
+    }
+
+    static com.concordium.sdk.responses.accountinfo.Cooldown to(Cooldown cooldown) {
+        return com.concordium.sdk.responses.accountinfo.Cooldown.builder()
+                .amount(to(cooldown.getAmount()))
+                .endTime(Timestamp.from(cooldown.getEndTime()))
+                .status(com.concordium.sdk.responses.accountinfo.Cooldown.CooldownStatus.from(cooldown.getStatus()))
+                .build();
     }
 
     @Nullable
@@ -1232,14 +1243,24 @@ interface ClientV2MapperExtensions {
         return BakerPoolStatus.builder()
                 .bakerId(to(grpcOutput.getBaker()))
                 .bakerAddress(to(grpcOutput.getAddress()))
-                .bakerEquityCapital(to(grpcOutput.getEquityCapital()))
-                .delegatedCapital(to(grpcOutput.getDelegatedCapital()))
-                .delegatedCapitalCap(to(grpcOutput.getDelegatedCapitalCap()))
-                .poolInfo(to(grpcOutput.getPoolInfo()))
+                .bakerEquityCapital(grpcOutput.hasEquityCapital()
+                        ? to(grpcOutput.getEquityCapital())
+                        : null)
+                .delegatedCapital(grpcOutput.hasDelegatedCapital()
+                        ? to(grpcOutput.getDelegatedCapital())
+                        : null)
+                .delegatedCapitalCap(grpcOutput.hasDelegatedCapitalCap()
+                        ? to(grpcOutput.getDelegatedCapitalCap())
+                        : null)
+                .poolInfo(grpcOutput.hasPoolInfo()
+                        ? to(grpcOutput.getPoolInfo())
+                        : null)
                 .bakerStakePendingChange(grpcOutput.hasEquityPendingChange()
                         ? to(grpcOutput.getEquityPendingChange())
                         : null)
-                .currentPaydayStatus(grpcOutput.hasCurrentPaydayInfo() ? to(grpcOutput.getCurrentPaydayInfo()) : null)
+                .currentPaydayStatus(grpcOutput.hasCurrentPaydayInfo()
+                        ? to(grpcOutput.getCurrentPaydayInfo())
+                        : null)
                 .allPoolTotalCapital(to(grpcOutput.getAllPoolTotalCapital()))
                 .build();
     }