+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/transactionstatus/AbstractBakerResult.java
@@ -1,6 +1,5 @@
 package com.concordium.sdk.responses.transactionstatus;
 
-import com.concordium.sdk.responses.BakerId;
 import com.concordium.sdk.types.AccountAddress;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -12,7 +11,5 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 @EqualsAndHashCode
 public abstract class AbstractBakerResult implements TransactionResultEvent {
-    private final BakerId bakerId;
     private final AccountAddress account;
-
 }