+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+- Support for Protocol 7
+- Added `ProtocolVersion.V7` corresponding to Protocol version 7
+- Added `cooldowns` list to `AccountInfo`
+- Added `availableBalance` to `AccountInfo`
+- Made optional the following `BakerPoolStatus` fields: 
+  `bakerEquityCapital`, `delegatedCapital`, `delegatedCapitalCap`, `poolInfo`
+- Added `BakerDelegationRemoved` and `DelegationBakerRemoved` transaction result events
+
 ## 7.2.0
 - Added `MessageSigningDigest` class to generate digests for message signing
 - Added support for company identity attributes